ticket_issuance_system/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
hiredis==2.2.3
httpie==3.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.2.0
openpyxl==3.1.2
pdfkit==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wkhtmltopdf==0.2
zope.interface==6.0
